Change cost function to yield better truncation behaviour over a wide range of string lengths
If you requested a truncation to a quite short string length (say 10
characters or so) in the the old way of calculating, the truncation
algorithm would often just extend to the end of the string, even if the
input string was 20 characters long. I've changed the costs to be in
terms of proportion of target string length, rather than in terms of
absolute characters, which allows the truncator to operate well over a
large range of different string lengths.